// frontend/src/api/generated/model/reviewFinding.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewFindingFilePath } from './reviewFindingFilePath';
import type { ReviewFindingLineEnd } from './reviewFindingLineEnd';
import type { ReviewFindingLineStart } from './reviewFindingLineStart';
import type { FindingSeverity } from './findingSeverity';
import type { FindingStatus } from './findingStatus';

/**
 * A single review finding
 */
export interface ReviewFinding {
  description: string;
  file_path?: ReviewFindingFilePath;
  id: string;
  line_end?: ReviewFindingLineEnd;
  line_start?: ReviewFindingLineStart;
  severity: FindingSeverity;
  status: FindingStatus;
  title: string;
}