// frontend/src/api/generated/model/findingStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenCode Studio API
 * API for OpenCode Studio - AI-powered development platform
 * OpenAPI spec version: 0.1.0
 */

/**
 * Status of a finding
 */
export type FindingStatus = typeof FindingStatus[keyof typeof FindingStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FindingStatus = {
  pending: 'pending',
  fixed: 'fixed',
  skipped: 'skipped',
} as const;